LaatUI/requirements.txt
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
amqp==5.1.0
anyjson==0.3.3
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
celery-with-redis==3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==36.0.2
Deprecated==1.2.13
Django==3.2.12
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-cors-headers==3.10.1
django-filter==21.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
drf-yasg==1.20.0
flower==1.0.0
gunicorn==20.1.0
humanize==3.14.0
idna==3.3
importlib-metadata==4.8.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
kombu==5.1.0
Markdown==3.3.6
MarkupSafe==2.0.1
oss2==2.15.0
packaging==21.3
prometheus-client==0.13.1
prompt-toolkit==3.0.28
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.11.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.2.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
types-mysqlclient==2.0.5
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.6.0
